requirements.txt
bokeh==3.0.3
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.24.0
pandas==1.5.2
Pyrebase4==4.6.0
requests==2.28.1
